validate_result/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
cachetools==5.5.2
cart==1.2.2
certifi==2024.12.14
charset-normalizer==3.4.1
deepseek==1.0.0
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-ckeditor==6.7.2
django-js-asset==3.0.1
dotenv==0.9.9
eval_type_backport==0.2.2
filelock==3.16.1
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
jiter==0.8.2
jsonpath-python==1.0.6
mistralai==1.5.0
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.65.2
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
selenium==4.27.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
websocket-client==1.8.0
wsproto==1.2.0
